Add control modifiers
